services/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
redis==6.4.0
httpx==0.28.1
PyJWT==2.10.1
twilio==9.8.0
python-json-logger==3.3.0
jinja2==3.1.6
